feat: add a name for notifications
