+++ react-chat/src/hooks/useCentrifuge.js
@@ -6,27 +6,7 @@ import {
   getSubscribeToken,
 } from '../api/centrifugo/index.js';
 import EVENT_CENTRIFUGO from '../const/events.js';
-
-const sendNotification = (message) => {
-  if (!('Notification' in window)) {
-    console.warn('Этот браузер не поддерживает уведомления на рабочем столе');
-    return;
-  }
-
-  if (Notification.permission === 'granted') {
-    new Notification('Новое сообщение', {
-      body: message.text,
-    });
-  } else if (Notification.permission !== 'denied') {
-    Notification.requestPermission().then((permission) => {
-      if (permission === 'granted') {
-        new Notification('Новое сообщение', {
-          body: message.text,
-        });
-      }
-    });
-  }
-};
+import sendNotification from '../helpers/sendNotification.js';
 
 const useCentrifuge = (chatId, setMessages) => {
   useEffect(() => {
@@ -54,6 +34,8 @@ const useCentrifuge = (chatId, setMessages) => {
           const { event, message } = ctx.data;
 
           const isCurrentChat = message.chat === chatId;
+
+          console.log(message);
 
           if (!isCurrentChat) {
             sendNotification(message);
